Cut over to an entirely new implementation.
This new implementation uses a more object-oriented approach to
validationer. Now, each function must return either true, false, or a
ValidationResult of either [ValidationSuccess, ValidationError].

These objects encapsulate everything that is needed for post-processing
logic later on. For now, they just include properties such as
"errorMessage" or "isSuccess".

This commit also splits the project up into a src folder. Below is a
brief description of what each of the files is going to be doing.

---
src/validations/coreValidations.js - This will probably be removed in
favor of single-issue-files in the future. For now, it's a testing bed
for simple validations.

-

src/engine.js - Right now, this contains all the logic that turns an
object and a set of validations into a "validation structure". That is,
an object built up of ValidationResult objects.

-

src/rule.js - Commented out for now. The idea is that eventually, you'll
be able to make classes which act as validation rules. However, I was
raised as a functional programmer so I don't see immediate value here.

-

src/validationResults.js - This contains the class definitions for the
ValidationResult base class, as well as ValidationSuccess and
ValidationError. These are exposed to functions used in validationer as
functional constructors, taking only an exception message.