Add Unit Tests which test all validations
This adds a large unit test which runs every known validation against a
data structure designed to be 100% true. This unit test found a few bugs
during its first few runs, so those bugs are also patched in this
commit.

---

This is one of the last commits that will be issued directly to master.
Once a valid README, LICENSE, and distillation method have been added to
the project, the master branch will be locked. For a semi-official TODO
for the project:

* TODO validationer [/]
- [ ] Distillation method without options
- [ ] Distillation unit test
- [ ] Full code documentation
- [ ] README.md
- [ ] LICENSE
- [ ] Github project board

Then, TODO's will be taken up with the project board.