Wrap the validationerEngine in simpleDistill
simpleDistill is the beginning of the distillation module for
validationer. It takes a "validation structure" (the output of the
validationer engine) and makes it all into a single boolean.

This means that you can use it to make simple boolean tests from largely
complex testing conditions. In the future, this will be split up more
than it currently is. The engine won't call a distiller directly.

You will be able to inspect the validation structure independant of the
distillation value, since a lot of useful debugging values are lost with
it.